# public/swagger/collection/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Negotium Solutions Swagger API Documentation",
        "description": "This is a Negotium swagger API documentation. You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,#swagger](http://swagger.io/irc/).For this sample, you can use the api key `special-key` to test the authorization filters",
        "contact": {
            "name": "Admin",
            "email": "admin@negotium-solutions.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.negotium-solutions.com/api",
            "description": "This is the Negotium Solutions Swagger API Server"
        }
    ],
    "paths": {
        "/{tenant}/activity/{schema_id}/{id}": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Get a activity",
                "description": "Authenticate using a bearer token",
                "operationId": "getActivity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Activity Id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/{tenant}/activity/{schema_id}": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Get Activities",
                "description": "Authenticate using a bearer token",
                "operationId": "getActivities",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/{tenant}/activity/create/{schema_id}": {
            "post": {
                "tags": [
                    "Activity"
                ],
                "summary": "Create a new activity",
                "description": "Authenticate using a bearer token",
                "operationId": "createActivity",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "process_step_id": 2,
                                    "name": "activity",
                                    "columns": [
                                        {
                                            "name": "name",
                                            "type": "string",
                                            "attributes": {
                                                "required": 1,
                                                "indent": 0,
                                                "is_id": 0,
                                                "is_passport": 0
                                            }
                                        },
                                        {
                                            "name": "email",
                                            "type": "string",
                                            "attributes": {
                                                "required": 1,
                                                "indent": 0,
                                                "is_id": 0,
                                                "is_passport": 0
                                            }
                                        },
                                        {
                                            "name": "age",
                                            "type": "integer",
                                            "attributes": {
                                                "required": 1,
                                                "indent": 0,
                                                "is_id": 0,
                                                "is_passport": 0
                                            }
                                        },
                                        {
                                            "name": "biography",
                                            "type": "text",
                                            "attributes": {
                                                "required": 1,
                                                "indent": 0,
                                                "is_id": 0,
                                                "is_passport": 0
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "code": 201,
                                                "message": "Activity successfully created.",
                                                "data": {
                                                    "id": 7
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/{tenant}/activity/update/{id}": {
            "put": {
                "tags": [
                    "Activity"
                ],
                "summary": "Update a activity",
                "description": "Authenticate using a bearer token",
                "operationId": "updateActivity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Activity Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Input validation error"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/{tenant}/activity/delete/{id}": {
            "delete": {
                "tags": [
                    "Activity"
                ],
                "summary": "Delete a activity by ID.",
                "description": "Authenticate using a bearer token",
                "operationId": "deleteActivityById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}